Complete trivial animations of movement
