+++ src/arena/battle_scene.rs
@@ -260,41 +260,3 @@ fn is_keystroke_skill(keycode: Keycode) -> Option<u32> {
         None
     }
 }
-
-pub fn tick_animations(ecs: &World, frame: u64) -> BoxResult<()> {
-    // Remove completed animations, applying their change
-    let mut completed = vec![];
-    let mut to_move = vec![];
-    {
-        let entities = ecs.read_resource::<specs::world::EntitiesRes>();
-        let animations = ecs.read_storage::<AnimationComponent>();
-        let mut positions = ecs.write_storage::<PositionComponent>();
-
-        for (entity, animation, position) in (&entities, &animations, (&mut positions).maybe()).join() {
-            if animation.is_complete(frame) {
-                completed.push(entity);
-            }
-            if let Animation::Position { start: _, end } = &animation.animation {
-                if let Some(_) = position {
-                    to_move.push((entity, *end));
-                }
-            }
-        }
-    }
-    for (entity, position) in to_move.iter() {
-        // This code needs to live outside of apply_move (should be renamed)
-        // Like start_move and complete_move?
-        //let mut positions = ecs.write_storage::<PositionComponent>();
-        //let position = &mut positions.get_mut(*entity).unwrap();
-        //position.move_to(*new_position);
-        apply_move(&ecs, &entity, position);
-    }
-    {
-        let mut animations = ecs.write_storage::<AnimationComponent>();
-        for c in completed {
-            animations.remove(c);
-        }
-    }
-
-    Ok(())
-}
